Adding my game jam stuffs! - Nick
Find the correct spot to stand and the (roughly) correct orientation in
order to make the projected image along the geometry look like the
picture in the bottom-right!

I’ve thrown my project under the SpotFinder folder in the root. The
HallwayProjection is the primary scene. Oughta be playable with
keyboard/mouse or gamepad, but there could possibly be a few tricky or
impossible rotations chosen by the randomization code.